backend/requirements.txt
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
alembic>=1.12.0
redis>=5.0.0

# Neo4j
neo4j>=5.14.0

# Tree-sitter parsing
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4
tree-sitter-javascript>=0.20.4
tree-sitter-typescript>=0.20.4
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.4
tree-sitter-java>=0.20.2
tree-sitter-c>=0.20.6
tree-sitter-cpp>=0.20.5

# Embeddings and ML
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# MCP Protocol
mcp>=1.0.0

# Utilities
httpx>=0.25.0
GitPython>=3.1.40
python-dotenv>=1.0.0
structlog>=23.2.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
